radbc_safeguard: Added a bug fix for a potentially serious issue with read_obs.F90 found by Anna Shlyaeva. The bug occurs when setting up locations of first processor to read observations of a given type. The code fails to assign processor values to the observations when npestart < npe and all the remaining non-zero values in npe_sub3 are greater than npe-npestart. Observations for which the above conditions are satisfied will not be read in GSI.
Both a quick fix and an alternative method for setting up the location of the first processor were given and it was decided to move forward with the alternative method. This alternative method won't give the same distribution as the previous code in all cases, but it's simpler and easier to read.

The regression tests (both standard and debug) were run and all returned bit-reproducible results.