Changes required for building and running the HWRF gsi on Hera.
Created new Makefile.conf.hera files.
Updated to use the Intel Compiler 18 family.
Update to use more recent NCEPDEV/nwprod/NCEPLIBS
Updated CRTM library to 2.2.6 from 2.2.3.

Fixes a Hera runtime error, where the
only code modifications were in the control2state files
to initialize pointers that were causing a runtime error
when compiled with Optimization level 3 -O3.

As a note: Added  Makefile.conf.hera.moduleload files.
These are not being used, but can be, if the libraries that
are required to build GSI are loaded with the module command.